+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 1.7.0 - 2020-11-23
+## Added
+- Support for [Addin-only](https://www.chrishasz.com/spgo/authentication/addin-only-authentication) authentication, so that you have another work-around for MFA.
+### Changed
+- Improved error logging to add more exception detail to the extension output window.
+
 ## 1.6.3 - 2020-04-15
 ### Fixed
 - I found (and fixed) another corner-case for [This issue](https://github.com/chrishasz/spgo/issues/112).